feat(pick): move pickers to extras telescope & fzf-lua and include telescope by default
